Remove Factory, rename TestHandler -> ClassDefinitionFactory
